Added Profile Context, Code refactoring
